+++ src/VisualStudio/Core/Test.Next/UnifiedSettings/TestModel/UnifiedSettingBase.cs
@@ -28,113 +28,5 @@ namespace Roslyn.VisualStudio.Next.UnitTests.UnifiedSettings.TestModel
 
         [JsonPropertyName("migration")]
         public required Migration Migration { get; init; }
-
-        public static UnifiedSettingsOption<T> CreateOption<T>(
-            IOption2 onboardedOption,
-            string title,
-            int order,
-            T? defaultValue = default,
-            (IOption2 featureFlagOption, T value)? featureFlagAndExperimentValue = null,
-            (IOption2 enableWhenOption, object whenValue)? enableWhenOptionAndValue = null,
-            string? languageName = null) where T : notnull
-        {
-            var migration = new Migration { Pass = new Pass { Input = new Input(onboardedOption, languageName) } };
-            var type = onboardedOption.Definition.Type;
-            // If the option's type is nullable type, like bool?, we use bool in the registration file.
-            var underlyingType = Nullable.GetUnderlyingType(type);
-            if (underlyingType is not null)
-            {
-                Assert.True(featureFlagAndExperimentValue is not null);
-            }
-            var nonNullableType = underlyingType ?? type;
-
-            var alternativeDefault = featureFlagAndExperimentValue is not null
-                ? new AlternativeDefault<T>(featureFlagAndExperimentValue.Value.featureFlagOption, featureFlagAndExperimentValue.Value.value)
-                : null;
-
-            var enableWhen = enableWhenOptionAndValue is not null
-                ? $"config:{UnifiedSettingsTests.s_optionToUnifiedSettingPath[enableWhenOptionAndValue.Value.enableWhenOption]}='{enableWhenOptionAndValue.Value.whenValue}'"
-                : null;
-
-            var expectedDefault = defaultValue ?? onboardedOption.Definition.DefaultValue;
-            // If the option default value is null, it means the option is in experiment mode and is hidden by a feature flag.
-            // In Unified Settings it is not allowed and should be replaced by using the alternative default.
-            // Like:
-            //     "textEditor.csharp.intellisense.showNewSnippetExperience": {
-            //         "type": "boolean",
-            //         "default": false,
-            //         "alternateDefault": {
-            //             "flagName": "Roslyn.SnippetCompletion",
-            //             "default": true
-            //         }
-            //      }
-            // so please specify a non-null default value.
-            Assert.NotNull(expectedDefault);
-
-            return new UnifiedSettingsOption<T>
-            {
-                Title = title,
-                Type = nonNullableType.Name.ToCamelCase(),
-                Order = order,
-                EnableWhen = enableWhen,
-                Migration = migration,
-                AlternativeDefault = alternativeDefault,
-                Default = (T)expectedDefault
-            };
-        }
-
-        public static UnifiedSettingsEnumOption CreateEnumOption<T>(
-            IOption2 onboardedOption,
-            string title,
-            int order,
-            T? defaultValue,
-            string[] enumLabels,
-            T[]? enumValues = null,
-            (IOption2 featureFlagOption, T value)? featureFlagAndExperimentValue = null,
-            (IOption2 enableWhenOption, object whenValue)? enableWhenOptionAndValue = null,
-            string? languageName = null) where T : Enum
-        {
-            var type = onboardedOption.Definition.Type;
-            // If the option's type is nullable type, we use the original type in the registration file.
-            var nonNullableType = Nullable.GetUnderlyingType(type) ?? type;
-            Assert.Equal(typeof(T), nonNullableType);
-
-            var expectedEnumValues = enumValues ?? [.. Enum.GetValues(nonNullableType).Cast<T>()];
-            var migration = new Migration
-            {
-                EnumToInteger = new EnumToInteger
-                {
-                    Input = new Input(onboardedOption, languageName),
-                    Map = new Map()
-                    {
-                        EnumValueMatches = [.. expectedEnumValues.Select(value => new Map.EnumToValuePair { Result = value.ToString().ToCamelCase(), Match = Convert.ToInt32(value) })]
-                    }
-                }
-            };
-
-            var alternativeDefault = featureFlagAndExperimentValue is not null
-                ? new AlternativeDefault<string>(featureFlagAndExperimentValue.Value.featureFlagOption, featureFlagAndExperimentValue.Value.value.ToString().ToCamelCase())
-                : null;
-
-            var enableWhen = enableWhenOptionAndValue is not null
-                ? $"config:{UnifiedSettingsTests.s_optionToUnifiedSettingPath[enableWhenOptionAndValue.Value.enableWhenOption]}='{enableWhenOptionAndValue.Value.whenValue}'"
-                : null;
-
-            var expectedDefault = defaultValue ?? onboardedOption.Definition.DefaultValue;
-            Assert.NotNull(expectedDefault);
-
-            return new UnifiedSettingsEnumOption()
-            {
-                Title = title,
-                Type = "string",
-                Enum = [.. expectedEnumValues.Select(value => value.ToString())],
-                EnumLabel = enumLabels,
-                Order = order,
-                EnableWhen = enableWhen,
-                Migration = migration,
-                AlternativeDefault = alternativeDefault,
-                Default = expectedDefault.ToString().ToCamelCase(),
-            };
-        }
     }
 }
